webui: Split into multiple, smaller components
